requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
faiss-cpu==1.11.0.post1
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.3.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
tiktoken==0.10.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
